// frontend/src/app/api/model/area.ts
/**
 * Reservation System API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Area { 
    id: string;
    name: string;
    color: string;
    maxBookingDurationMinutes: number;
    maxBookingEndOffsetDays?: number;
    sortOrder?: number;
}