// frontend/src/app/api/model/workplace.ts
/**
 * Reservation System API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Workplace { 
    id: string;
    name: string;
    description?: string;
    status: Workplace.StatusEnum;
    location?: string;
    areaId: string;
    wikiUrl?: string;
    maxBookingDurationMinutes?: number | null;
    blocksWorkplaceIds?: Array<string>;
    blocksWorkplacesWithTag?: Array<string>;
    tags?: Array<string>;
    sortOrder?: number;
}
export namespace Workplace {
    export const StatusEnum = {
        Ok: 'OK',
        Defect: 'DEFECT',
        Disabled: 'DISABLED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}