// frontend/src/app/api/model/booking.ts
/**
 * Reservation System API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Booking { 
    id: string;
    creatorRole: string;
    workplaceId: string;
    /**
     * Only part of the read response when the booking blocks other workplaces
     */
    blocksWorkplaceIds?: Array<string>;
    /**
     * Visible only with viewBookingsDetails permission
     */
    name: string;
    /**
     * Visible only with viewBookingsDetails permission
     */
    contact: string;
    startTime: string;
    endTime: string;
    bookingSeriesId: string | null;
}